Anonymise factor levels

Replaces factor levels with arbitary numeric identifiers. Neither the values nor the order of the levels are preserved.

fct_anon(f, prefix = "")

Arguments

f

A factor

prefix

A character prefix to insert in front of the random labels.

Examples

gss_cat$relig %>% fct_count()

#> # A tibble: 16 × 2
#>                         f     n
#>                    <fctr> <int>
#> 1               No answer    93
#> 2              Don't know    15
#> 3 Inter-nondenominational   109
#> 4         Native american    23
#> 5               Christian   689
#> 6      Orthodox-christian    95
#> # ... with 10 more rows

gss_cat$relig %>% fct_anon() %>% fct_count()

#> # A tibble: 16 × 2
#>        f     n
#>   <fctr> <int>
#> 1     01   689
#> 2     02    93
#> 3     03    23
#> 4     04  5124
#> 5     05   104
#> 6     06    95
#> # ... with 10 more rows

gss_cat$relig %>% fct_anon() %>% fct_count()

#> # A tibble: 16 × 2
#>        f     n
#>   <fctr> <int>
#> 1     01    15
#> 2     02    32
#> 3     03  3523
#> 4     04   689
#> 5     05   388
#> 6     06   109
#> # ... with 10 more rows

gss_cat$relig %>% fct_anon("X") %>% fct_count()

#> # A tibble: 16 × 2
#>        f     n
#>   <fctr> <int>
#> 1    X01   388
#> 2    X02   104
#> 3    X03  5124
#> 4    X04   147
#> 5    X05    93
#> 6    X06    23
#> # ... with 10 more rows