Anonymise factor levels

Replaces factor levels with arbitrary numeric identifiers. Neither the values nor the order of the levels are preserved.

Usage

fct_anon(f, prefix = "")

Arguments

f

A factor.

prefix

A character prefix to insert in front of the random labels.

Examples

gss_cat$relig %>% fct_count()
#> # A tibble: 16 × 2
#>    f                           n
#>    <fct>                   <int>
#>  1 No answer                  93
#>  2 Don't know                 15
#>  3 Inter-nondenominational   109
#>  4 Native american            23
#>  5 Christian                 689
#>  6 Orthodox-christian         95
#>  7 Moslem/islam              104
#>  8 Other eastern              32
#>  9 Hinduism                   71
#> 10 Buddhism                  147
#> 11 Other                     224
#> 12 None                     3523
#> 13 Jewish                    388
#> 14 Catholic                 5124
#> 15 Protestant              10846
#> 16 Not applicable              0
gss_cat$relig %>%
  fct_anon() %>%
  fct_count()
#> # A tibble: 16 × 2
#>    f         n
#>    <fct> <int>
#>  1 01      689
#>  2 02      104
#>  3 03      224
#>  4 04       23
#>  5 05       71
#>  6 06        0
#>  7 07       15
#>  8 08      147
#>  9 09     5124
#> 10 10      109
#> 11 11       95
#> 12 12       32
#> 13 13     3523
#> 14 14       93
#> 15 15      388
#> 16 16    10846
gss_cat$relig %>%
  fct_anon("X") %>%
  fct_count()
#> # A tibble: 16 × 2
#>    f         n
#>    <fct> <int>
#>  1 X01     224
#>  2 X02    5124
#>  3 X03    3523
#>  4 X04   10846
#>  5 X05     147
#>  6 X06     388
#>  7 X07     104
#>  8 X08      93
#>  9 X09      15
#> 10 X10       0
#> 11 X11      71
#> 12 X12      95
#> 13 X13      32
#> 14 X14     109
#> 15 X15     689
#> 16 X16      23