Count entries in a factor

Count entries in a factor

fct_count(f, sort = FALSE, prop = FALSE)

Arguments

f
A factor (or character vector).

sort
If TRUE, sort the result so that the most common values float to the top.

prop
If TRUE, compute the fraction of marginal table.

Value

A tibble with columns f, n and p, if prop is TRUE.

Examples

f <- factor(sample(letters)[rpois(1000, 10)])
table(f)

#> f
#>   a   b   c   d   e   f   g   h   j   k   l   n   p   q   r   s   v   w   z 
#>  51 132 118  25  11  43  94   6  64 122  53  14 108  50   2 101   4   1   1 

fct_count(f)

#> # A tibble: 19 x 2
#>    f         n
#>    <fct> <int>
#>  1 a        51
#>  2 b       132
#>  3 c       118
#>  4 d        25
#>  5 e        11
#>  6 f        43
#>  7 g        94
#>  8 h         6
#>  9 j        64
#> 10 k       122
#> 11 l        53
#> 12 n        14
#> 13 p       108
#> 14 q        50
#> 15 r         2
#> 16 s       101
#> 17 v         4
#> 18 w         1
#> 19 z         1

fct_count(f, sort = TRUE)

#> # A tibble: 19 x 2
#>    f         n
#>    <fct> <int>
#>  1 b       132
#>  2 k       122
#>  3 c       118
#>  4 p       108
#>  5 s       101
#>  6 g        94
#>  7 j        64
#>  8 l        53
#>  9 a        51
#> 10 q        50
#> 11 f        43
#> 12 d        25
#> 13 n        14
#> 14 e        11
#> 15 h         6
#> 16 v         4
#> 17 r         2
#> 18 w         1
#> 19 z         1

fct_count(f, sort = TRUE, prop = TRUE)

#> # A tibble: 19 x 3
#>    f         n     p
#>    <fct> <int> <dbl>
#>  1 b       132 0.132
#>  2 k       122 0.122
#>  3 c       118 0.118
#>  4 p       108 0.108
#>  5 s       101 0.101
#>  6 g        94 0.094
#>  7 j        64 0.064
#>  8 l        53 0.053
#>  9 a        51 0.051
#> 10 q        50 0.05 
#> 11 f        43 0.043
#> 12 d        25 0.025
#> 13 n        14 0.014
#> 14 e        11 0.011
#> 15 h         6 0.006
#> 16 v         4 0.004
#> 17 r         2 0.002
#> 18 w         1 0.001
#> 19 z         1 0.001