fct_explicit_na {forcats}	R Documentation

Make missing values explicit

Description

This function is deprecated because the terminology is confusing; please use fct_na_value_to_level() instead.

This gives missing values an explicit factor level, ensuring that they appear in summaries and on plots.

Usage

fct_explicit_na(f, na_level = "(Missing)")

Arguments

f
A factor (or character vector).

na_level
Level to use for missing values: this is what NAs will be changed to.

Examples

f1 <- factor(c("a", "a", NA, NA, "a", "b", NA, "c", "a", "c", "b"))
fct_count(f1)
table(f1)
sum(is.na(f1))

# previously
f2 <- fct_explicit_na(f1)
# now
f2 <- fct_na_value_to_level(f1)

fct_count(f2)
table(f2)
sum(is.na(f2))

[Package forcats version 1.0.0 Index]