Lump together factor levels into "other"

fct_lump() lumps together least/most common levels (as defined by n and prop) into "other". fct_lump_min() lumps together all levels which don't appear at least min number of times.

fct_lump(
  f,
  n,
  prop,
  w = NULL,
  other_level = "Other",
  ties.method = c("min", "average", "first", "last", "random", "max")
)

fct_lump_min(f, min, w = NULL, other_level = "Other")

Arguments

f
A factor (or character vector).

n, prop
If both n and prop are missing, fct_lump() lumps together the least frequent levels into "other", while ensuring that "other" is still the smallest level. It's particularly useful in conjunction with fct_inorder().

Positive n preserves the most common n values. Negative n preserves the least common -n values. It there are ties, you will get at least abs(n) values.

Positive prop preserves values that appear at least prop of the time. Negative prop preserves values that appear at most -prop of the time.

w
An optional numeric vector giving weights for frequency of each value (not level) in f.

other_level
Value of level used for "other" values. Always placed at end of levels.

ties.method
A character string specifying how ties are treated. See rank() for details.

min
Preserves values that appear at least min number of times.

See also

Examples

x <- factor(rep(LETTERS[1:9], times = c(40, 10, 5, 27, 1, 1, 1, 1, 1)))
x %>% table()

#> .
#>  A  B  C  D  E  F  G  H  I 
#> 40 10  5 27  1  1  1  1  1 

x %>% fct_lump() %>% table()

#> .
#>     A     D Other 
#>    40    27    20 

x %>% fct_lump() %>% fct_inorder() %>% table()

#> .
#>     A Other     D 
#>    40    20    27 

x <- factor(letters[rpois(100, 5)])
x

#>   [1] f f e d d d c h f d g h f f d c b b f c h h e h h c f e f d b e c b f e c
#>  [38] h d b f e e f d c g f g d d e c a b d c g e f d f f g g e c f b g g b g e
#>  [75] g i e d g g d c d i d c b d c d d f f d e d h d d e
#> Levels: a b c d e f g h i

table(x)

#> x
#>  a  b  c  d  e  f  g  h  i 
#>  1  9 13 23 14 18 12  8  2 

table(fct_lump(x))

#> 
#>     b     c     d     e     f     g     h Other 
#>     9    13    23    14    18    12     8     3 

# Use positive values to collapse the rarest
fct_lump(x, n = 3)

#>   [1] f     f     e     d     d     d     Other Other f     d     Other Other
#>  [13] f     f     d     Other Other Other f     Other Other Other e     Other
#>  [25] Other Other f     e     f     d     Other e     Other Other f     e    
#>  [37] Other Other d     Other f     e     e     f     d     Other Other f    
#>  [49] Other d     d     e     Other Other Other d     Other Other e     f    
#>  [61] d     f     f     Other Other e     Other f     Other Other Other Other
#>  [73] Other e     Other Other e     d     Other Other d     Other d     Other
#>  [85] d     Other Other d     Other d     d     f     f     d     e     d    
#>  [97] Other d     d     e    
#> Levels: d e f Other

fct_lump(x, prop = 0.1)

#>   [1] f     f     e     d     d     d     c     Other f     d     g     Other
#>  [13] f     f     d     c     Other Other f     c     Other Other e     Other
#>  [25] Other c     f     e     f     d     Other e     c     Other f     e    
#>  [37] c     Other d     Other f     e     e     f     d     c     g     f    
#>  [49] g     d     d     e     c     Other Other d     c     g     e     f    
#>  [61] d     f     f     g     g     e     c     f     Other g     g     Other
#>  [73] g     e     g     Other e     d     g     g     d     c     d     Other
#>  [85] d     c     Other d     c     d     d     f     f     d     e     d    
#>  [97] Other d     d     e    
#> Levels: c d e f g Other

# Use negative values to collapse the most common
fct_lump(x, n = -3)

#>   [1] Other Other Other Other Other Other Other h     Other Other Other h    
#>  [13] Other Other Other Other Other Other Other Other h     h     Other h    
#>  [25] h     Other Other Other Other Other Other Other Other Other Other Other
#>  [37] Other h     Other Other Other Other Other Other Other Other Other Other
#>  [49] Other Other Other Other Other a     Other Other Other Other Other Other
#>  [61] Other Other Other Other Other Other Other Other Other Other Other Other
#>  [73] Other Other Other i     Other Other Other Other Other Other Other i    
#>  [85] Other Other Other Other Other Other Other Other Other Other Other Other
#>  [97] h     Other Other Other
#> Levels: a h i Other

fct_lump(x, prop = -0.1)

#>   [1] Other Other Other Other Other Other Other h     Other Other Other h    
#>  [13] Other Other Other Other b     b     Other Other h     h     Other h    
#>  [25] h     Other Other Other Other Other b     Other Other b     Other Other
#>  [37] Other h     Other b     Other Other Other Other Other Other Other Other
#>  [49] Other Other Other Other Other a     b     Other Other Other Other Other
#>  [61] Other Other Other Other Other Other Other Other b     Other Other b    
#>  [73] Other Other Other i     Other Other Other Other Other Other Other i    
#>  [85] Other Other b     Other Other Other Other Other Other Other Other Other
#>  [97] h     Other Other Other
#> Levels: a b h i Other

# Use weighted frequencies
w <- c(rep(2, 50), rep(1, 50))
fct_lump(x, n = 5, w = w)

#>   [1] f     f     e     d     d     d     c     h     f     d     g     h    
#>  [13] f     f     d     c     Other Other f     c     h     h     e     h    
#>  [25] h     c     f     e     f     d     Other e     c     Other f     e    
#>  [37] c     h     d     Other f     e     e     f     d     c     g     f    
#>  [49] g     d     d     e     c     Other Other d     c     g     e     f    
#>  [61] d     f     f     g     g     e     c     f     Other g     g     Other
#>  [73] g     e     g     Other e     d     g     g     d     c     d     Other
#>  [85] d     c     Other d     c     d     d     f     f     d     e     d    
#>  [97] h     d     d     e    
#> Levels: c d e f g h Other

# Use ties.method to control how tied factors are collapsed
fct_lump(x, n = 6)

#>   [1] f     f     e     d     d     d     c     Other f     d     g     Other
#>  [13] f     f     d     c     b     b     f     c     Other Other e     Other
#>  [25] Other c     f     e     f     d     b     e     c     b     f     e    
#>  [37] c     Other d     b     f     e     e     f     d     c     g     f    
#>  [49] g     d     d     e     c     Other b     d     c     g     e     f    
#>  [61] d     f     f     g     g     e     c     f     b     g     g     b    
#>  [73] g     e     g     Other e     d     g     g     d     c     d     Other
#>  [85] d     c     b     d     c     d     d     f     f     d     e     d    
#>  [97] Other d     d     e    
#> Levels: b c d e f g Other

fct_lump(x, n = 6, ties.method = "max")

#>   [1] f     f     e     d     d     d     c     Other f     d     g     Other
#>  [13] f     f     d     c     b     b     f     c     Other Other e     Other
#>  [25] Other c     f     e     f     d     b     e     c     b     f     e    
#>  [37] c     Other d     b     f     e     e     f     d     c     g     f    
#>  [49] g     d     d     e     c     Other b     d     c     g     e     f    
#>  [61] d     f     f     g     g     e     c     f     b     g     g     b    
#>  [73] g     e     g     Other e     d     g     g     d     c     d     Other
#>  [85] d     c     b     d     c     d     d     f     f     d     e     d    
#>  [97] Other d     d     e    
#> Levels: b c d e f g Other

x <- factor(letters[rpois(100, 5)])
fct_lump_min(x, min = 10)

#>   [1] f     f     Other f     e     g     d     f     f     e     c     f    
#>  [13] e     f     c     d     f     e     g     e     c     g     f     c    
#>  [25] d     c     c     f     e     c     d     c     Other e     Other Other
#>  [37] c     c     f     c     d     f     c     g     f     e     Other d    
#>  [49] Other c     Other e     d     e     d     Other g     Other f     e    
#>  [61] e     d     Other Other Other g     Other e     Other d     f     d    
#>  [73] g     f     d     c     Other e     f     d     f     d     c     d    
#>  [85] g     c     e     f     c     f     Other Other g     d     g     g    
#>  [97] f     c     c     d    
#> Levels: c d e f g Other