Lump together factor levels into "other"

A family for lumping together levels that meet some criteria.

fct_lump_min(): lumps levels that appear fewer than min times.

fct_lump_prop(): lumps levels that appear in fewer prop * n times.

fct_lump_n() lumps all levels except for the n most frequent (or least frequent if n < 0)

fct_lump_lowfreq() lumps together the least frequent levels, ensuring that "other" is still the smallest level.

fct_lump() exists primarily for historical reasons, as it automatically picks between these different methods depending on its arguments. We no longer recommend that you use it.

Usage

fct_lump(
  f,
  n,
  prop,
  w = NULL,
  other_level = "Other",
  ties.method = c("min", "average", "first", "last", "random", "max")
)

fct_lump_min(f, min, w = NULL, other_level = "Other")

fct_lump_prop(f, prop, w = NULL, other_level = "Other")

fct_lump_n(
  f,
  n,
  w = NULL,
  other_level = "Other",
  ties.method = c("min", "average", "first", "last", "random", "max")
)

fct_lump_lowfreq(f, other_level = "Other")

Arguments

f

A factor (or character vector).

n

Positive n preserves the most common n values. Negative n preserves the least common -n values. It there are ties, you will get at least abs(n) values.

prop

Positive prop lumps values which do not appear at least prop of the time. Negative prop lumps values that do not appear at most -prop of the time.

w

An optional numeric vector giving weights for frequency of each value (not level) in f.

other_level

Value of level used for "other" values. Always placed at end of levels.

ties.method

A character string specifying how ties are treated. See rank() for details.

min

Preserve levels that appear at least min number of times.

See also

Examples

x <- factor(rep(LETTERS[1:9], times = c(40, 10, 5, 27, 1, 1, 1, 1, 1)))
x %>% table()
#> .
#>  A  B  C  D  E  F  G  H  I 
#> 40 10  5 27  1  1  1  1  1 
x %>%
  fct_lump_n(3) %>%
  table()
#> .
#>     A     B     D Other 
#>    40    10    27    10 
x %>%
  fct_lump_prop(0.10) %>%
  table()
#> .
#>     A     B     D Other 
#>    40    10    27    10 
x %>%
  fct_lump_min(5) %>%
  table()
#> .
#>     A     B     C     D Other 
#>    40    10     5    27     5 
x %>%
  fct_lump_lowfreq() %>%
  table()
#> .
#>     A     D Other 
#>    40    27    20 

x <- factor(letters[rpois(100, 5)])
x
#>  [1] h e f b d e e e d c d b f g i e g b e c h c i g h e e f c a e e e f f
#> [36] e a k c g d g b b h d c f g h e d b d f c g h e c e b f f g f e f g j
#> [71] a d f f e d e d b e f g e h f d h d e j i c f e c e f d
#> Levels: a b c d e f g h i j k
table(x)
#> x
#>  a  b  c  d  e  f  g  h  i  j  k 
#>  3  8 10 13 23 17 10  8  3  2  1 
table(fct_lump_lowfreq(x))
#> 
#>     a     b     c     d     e     f     g     h     i     j Other 
#>     3     8    10    13    23    17    10     8     3     2     1 

# Use positive values to collapse the rarest
fct_lump_n(x, n = 3)
#>  [1] Other e     f     Other d     e     e     e     d     Other d    
#> [12] Other f     Other Other e     Other Other e     Other Other Other
#> [23] Other Other Other e     e     f     Other Other e     e     e    
#> [34] f     f     e     Other Other Other Other d     Other Other Other
#> [45] Other d     Other f     Other Other e     d     Other d     f    
#> [56] Other Other Other e     Other e     Other f     f     Other f    
#> [67] e     f     Other Other Other d     f     f     e     d     e    
#> [78] d     Other e     f     Other e     Other f     d     Other d    
#> [89] e     Other Other Other f     e     Other e     f     d    
#> Levels: d e f Other
fct_lump_prop(x, prop = 0.1)
#>  [1] Other e     f     Other d     e     e     e     d     c     d    
#> [12] Other f     g     Other e     g     Other e     c     Other c    
#> [23] Other g     Other e     e     f     c     Other e     e     e    
#> [34] f     f     e     Other Other c     g     d     g     Other Other
#> [45] Other d     c     f     g     Other e     d     Other d     f    
#> [56] c     g     Other e     c     e     Other f     f     g     f    
#> [67] e     f     g     Other Other d     f     f     e     d     e    
#> [78] d     Other e     f     g     e     Other f     d     Other d    
#> [89] e     Other Other c     f     e     c     e     f     d    
#> Levels: c d e f g Other

# Use negative values to collapse the most common
fct_lump_n(x, n = -3)
#>  [1] Other Other Other Other Other Other Other Other Other Other Other
#> [12] Other Other Other i     Other Other Other Other Other Other Other
#> [23] i     Other Other Other Other Other Other a     Other Other Other
#> [34] Other Other Other a     k     Other Other Other Other Other Other
#> [45] Other Other Other Other Other Other Other Other Other Other Other
#> [56] Other Other Other Other Other Other Other Other Other Other Other
#> [67] Other Other Other j     a     Other Other Other Other Other Other
#> [78] Other Other Other Other Other Other Other Other Other Other Other
#> [89] Other j     i     Other Other Other Other Other Other Other
#> Levels: a i j k Other
fct_lump_prop(x, prop = -0.1)
#>  [1] h     Other Other b     Other Other Other Other Other Other Other
#> [12] b     Other Other i     Other Other b     Other Other h     Other
#> [23] i     Other h     Other Other Other Other a     Other Other Other
#> [34] Other Other Other a     k     Other Other Other Other b     b    
#> [45] h     Other Other Other Other h     Other Other b     Other Other
#> [56] Other Other h     Other Other Other b     Other Other Other Other
#> [67] Other Other Other j     a     Other Other Other Other Other Other
#> [78] Other b     Other Other Other Other h     Other Other h     Other
#> [89] Other j     i     Other Other Other Other Other Other Other
#> Levels: a b h i j k Other

# Use weighted frequencies
w <- c(rep(2, 50), rep(1, 50))
fct_lump_n(x, n = 5, w = w)
#> Error in fct_lump_n(x, n = 5, w = w): `w` must be the same length as `f` (98), not length 100

# Use ties.method to control how tied factors are collapsed
fct_lump_n(x, n = 6)
#>  [1] h     e     f     b     d     e     e     e     d     c     d    
#> [12] b     f     g     Other e     g     b     e     c     h     c    
#> [23] Other g     h     e     e     f     c     Other e     e     e    
#> [34] f     f     e     Other Other c     g     d     g     b     b    
#> [45] h     d     c     f     g     h     e     d     b     d     f    
#> [56] c     g     h     e     c     e     b     f     f     g     f    
#> [67] e     f     g     Other Other d     f     f     e     d     e    
#> [78] d     b     e     f     g     e     h     f     d     h     d    
#> [89] e     Other Other c     f     e     c     e     f     d    
#> Levels: b c d e f g h Other
fct_lump_n(x, n = 6, ties.method = "max")
#>  [1] Other e     f     Other d     e     e     e     d     c     d    
#> [12] Other f     g     Other e     g     Other e     c     Other c    
#> [23] Other g     Other e     e     f     c     Other e     e     e    
#> [34] f     f     e     Other Other c     g     d     g     Other Other
#> [45] Other d     c     f     g     Other e     d     Other d     f    
#> [56] c     g     Other e     c     e     Other f     f     g     f    
#> [67] e     f     g     Other Other d     f     f     e     d     e    
#> [78] d     Other e     f     g     e     Other f     d     Other d    
#> [89] e     Other Other c     f     e     c     e     f     d    
#> Levels: c d e f g Other

# Use fct_lump_min() to lump together all levels with fewer than `n` values
table(fct_lump_min(x, min = 10))
#> 
#>     c     d     e     f     g Other 
#>    10    13    23    17    10    25 
table(fct_lump_min(x, min = 15))
#> 
#>     e     f Other 
#>    23    17    58